cleaned up and made code more consistent
cleaned up the documentation in the ini files
added timeout for fieldtrip buffer in plot spectral and plot signal
(for consistency)
use integer number of samples in acceleration.py

+++ module/accelerometer/accelerometer.py
@@ -94,7 +94,7 @@ if debug>1:
     print hdr_input.labels
 
 # get the processing arameters
-window = patch.getfloat('processing','window')*hdr_input.fSample
+window = int(patch.getfloat('processing','window') * hdr_input.fSample) # expressed in samples
 
 channel_items = config.items('input')
 channel_name = []
